Adding the initial "BasicDetection" test scene for AI.
Created a "SpawnCube" script that allows you to create a volume filled with cubes for testing purposes.  Could be repurposed for volume painting (asteroids).

Added a new "Range" data class that allows you to specify a range of values and easily use operations to test the range with.

Updated some ship component parts for navigation, sensor, and an example ship.